+++ src/Constraints/ZipCode.php
@@ -2,6 +2,7 @@
 
 namespace SLLH\IsoCodesValidator\Constraints;
 
+use IsoCodes;
 use Symfony\Component\Validator\Constraint;
 use Symfony\Component\Validator\Exception\ConstraintDefinitionException;
 
@@ -13,13 +14,28 @@ use Symfony\Component\Validator\Exception\ConstraintDefinitionException;
  */
 class ZipCode extends Constraint
 {
+    /**
+     * @deprecated since 1.1, to be removed in 2.0.
+     */
     const US            = 'US';
+    /**
+     * @deprecated since 1.1, to be removed in 2.0.
+     */
     const CANADA        = 'Canada';
+    /**
+     * @deprecated since 1.1, to be removed in 2.0.
+     */
     const FRANCE        = 'France';
+    /**
+     * @deprecated since 1.1, to be removed in 2.0.
+     */
     const NETHERLANDS   = 'Netherlands';
 
     const ALL           = 'all';
 
+    /**
+     * @deprecated since 1.1, to be removed in 2.0.
+     */
     public static $countries = [
         self::US,
         self::CANADA,
@@ -38,8 +54,21 @@ class ZipCode extends Constraint
     {
         parent::__construct($options);
 
-        if ($this->country != self::ALL && !in_array($this->country, self::$countries)) {
-            throw new ConstraintDefinitionException(sprintf('The option "country" must be one of "%s" or "all"', implode('", "', self::$countries)));
+        // 1.0 BC
+        if ($this->country !== 'US' && in_array($this->country, self::$countries, true)) {
+            $deprecatedOptionsBridge = [
+                'Canada'        => 'CA',
+                'France'        => 'FR',
+                'Netherlands'   => 'NL',
+            ];
+
+            trigger_error('The value "'.$this->country.'" for '.__CLASS__.'::country options is deprecated since version 1.1 and will be removed in 2.0. Use '.$deprecatedOptionsBridge[$this->country].' instead.', E_USER_DEPRECATED);
+
+            $this->country = $deprecatedOptionsBridge[$this->country];
+        }
+
+        if ($this->country != self::ALL && !in_array($this->country, IsoCodes\ZipCode::getAvailableCountries())) {
+            throw new ConstraintDefinitionException(sprintf('The option "country" must be one of "%s" or "all"', implode('", "', IsoCodes\ZipCode::getAvailableCountries())));
         }
     }
 }