+++ tests/Constraints/ZipCodeTest.php
@@ -19,25 +19,57 @@ class ZipCodeTest extends \PHPUnit_Framework_TestCase
     public function getValidCountries()
     {
         return [
-            ['Canada'],
-            ['France'],
-            ['Netherlands'],
+            ['CA'],
+            ['FR'],
+            ['NL'],
             ['US'],
         ];
     }
 
     /**
+     * 1.0 BC.
+     *
+     * @dataProvider getLegacyValidCountries
+     */
+    public function testLegacyValidCountries($country)
+    {
+        $zipCode = new ZipCode(['country' => $country]);
+
+        // 1.0 BC
+        $deprecatedOptionsBridge = [
+            'Canada'        => 'CA',
+            'France'        => 'FR',
+            'Netherlands'   => 'NL',
+        ];
+
+        $this->assertSame($deprecatedOptionsBridge[$country], $zipCode->country);
+    }
+
+    public function getLegacyValidCountries()
+    {
+        return [
+            ['Canada'],
+            ['France'],
+            ['Netherlands'],
+        ];
+    }
+
+    /**
      * @dataProvider getInvalidCountries
-     * @expectedException Symfony\Component\Validator\Exception\ConstraintDefinitionException
+     * @expectedException \Symfony\Component\Validator\Exception\ConstraintDefinitionException
      */
     public function testInvalidCountries($country)
     {
-        $zipCode = new ZipCode(['country' => $country]);
+        new ZipCode(['country' => $country]);
     }
 
     public function getInvalidCountries()
     {
         return [
+            ['TL'],
+            [42],
+            ['Liberlands'],
+            // 1.0 BC Test
             ['canada'],
             ['CANADA'],
             ['caNada'],
@@ -47,6 +79,7 @@ class ZipCodeTest extends \PHPUnit_Framework_TestCase
             ['netherlands'],
             ['NETHERLANDS'],
             ['neTherlands'],
+            // END 1.0 BC Test
             ['us'],
             ['uS'],
             ['Us'],